Original matlab code
Pursuit:
Which physical components of sound must be manipulated for individual to perceive integration vs segregation of sound. Once this baseline is established, we will work with the sound threshold to understand whether visual stimulation masks this auditory perception (i.e. where the participant previously responded hearing segregated source of sound, it now hears integrated due to the visual stimulation).